Refactor paper_weight_agent.py into modular components
Split the 2039-line paper_weight_agent.py into smaller focused modules:
- paper_weight_models.py (355 lines): Data models (AssessmentDetail, DimensionScore, PaperWeightResult)
- paper_weight_extractors.py (448 lines): Pure functions for rule-based study type and sample size extraction
- paper_weight_llm_assessors.py (393 lines): LLM-based assessor functions for methodological quality and risk of bias
- paper_weight_prompts.py (182 lines): LLM prompt templates
- paper_weight_db.py (493 lines): Database operations for caching and persistence
- paper_weight_agent.py (485 lines): Main agent class orchestrating all components

All files now under 500 lines. Backward compatibility maintained through re-exports.
